+++ package/lib/lo_form.dart
@@ -7,15 +7,18 @@ export 'fields/lo_text_field.dart';
 // Core
 export 'src/lo_config.dart';
 export 'src/lo_field.dart';
+export 'src/lo_field_base_validator.dart';
 export 'src/lo_field_state.dart';
 export 'src/lo_form.dart';
+export 'src/lo_form_base_validator.dart';
 export 'src/lo_form_state.dart';
 export 'src/lo_status.dart';
-export 'src/lo_validator.dart';
 export 'src/types.dart';
 
 // Validators
-export 'validators/lo_custom_validator.dart';
-export 'validators/lo_length_validator.dart';
-export 'validators/lo_reg_exp_validator.dart';
-export 'validators/lo_required_validator.dart';
+export 'validators/field/lo_field_validator.dart';
+export 'validators/field/lo_length_validator.dart';
+export 'validators/field/lo_reg_exp_validator.dart';
+export 'validators/field/lo_required_validator.dart';
+export 'validators/form/lo_form_validator.dart';
+export 'validators/form/lo_match_validator.dart';