+++ package/lib/src/lo_field_base_validator.dart
@@ -1,11 +1,14 @@
-abstract class LoValidator<T> {
+abstract class LoFieldBaseValidator<T> {
   /// Returns `String` containing the error message if [value] is invalid.
   /// or `null` when [value] is valid.
   String? validate(T? value);
 
   /// Runs all [validators] on [value], then return the error message or
   /// `null` accordingly.
-  static String? reduce<T>(List<LoValidator> validators, T value) {
+  static String? run<T>(
+    List<LoFieldBaseValidator> validators,
+    T value,
+  ) {
     for (final validator in validators) {
       final error = validator.validate(value);
       if (error != null) return error;