+++ package/lib/src/types.dart
@@ -7,7 +7,6 @@ typedef FieldsMap = Map<String, LoFieldState>;
 typedef ValMap = Map<String, dynamic>;
 typedef ErrMap = Map<String, String?>;
 
-typedef ValidateFunc = ErrMap? Function(ValMap);
 typedef StatusCheckFunc = bool Function(LoStatus);
 typedef SetErrFunc = void Function(ErrMap);
 typedef SubmitFunc = FutureOr<bool?>? Function(ValMap, SetErrFunc);