+++ package/lib/validators/field/lo_field_validator.dart
@@ -1,0 +1,10 @@
+import '../../src/lo_field_base_validator.dart';
+
+class LoFieldValidator<T> implements LoFieldBaseValidator<T> {
+  final String? Function(T?) validateFunction;
+
+  LoFieldValidator(this.validateFunction);
+
+  @override
+  String? validate(T? value) => validateFunction(value);
+}